requirements.txt
aiohttp>=3.8.0
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
googlemaps==4.10.0